repository: implement repository locator
Zero or more repositories can be found on a given path when searching
up to a specified depth. The tests for RepositoryLocator are slightly
more integration-y than unit-y so move them to the tests directory.
The test cases themselves are unchanged. A new type RepoProbe has
been added in order to decouple traversal and repo identification.