repository: allow DI for repository type
Use the new mockable GitRepository trait for the repo locator. The type
GitRepoProbe was moved to the end because that is less likely to change
so it shouldn't need to be specified. This makes our types shorter. As
an added benefit, the GitRepositoryLocator can now be an alias instead
of a new type. This removes one layer of indirection and eliminates two
items from the public API.